Make live nginx redirect from www subdomain
Currently the ngnix config forces all requests to be made to www.lag.tv
via redirects.

Info from the currently used live certificate indicates that a bare SAN
is available (i.e. lag.tv is authorized along with www.lag.tv), and so
does not require a www subdomain.

This commit alters the nginx config to not rewrite requests to lag.tv,
and to rewrite anything else. This is no-www Class B.

Non-matching domains' rewrite target is changed to lag.tv.